# sdks/sdkwork-account-app-sdk/openapi/account-app-api.openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "SDKWork Account App API",
    "version": "3.0.0",
    "description": "Read-only wallet, billing, and account summary surfaces for authenticated app users.",
    "x-sdkwork-owner": "sdkwork-account",
    "x-sdkwork-api-authority": "sdkwork-account.app"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/app/v3/api/billing/history": {
      "get": {
        "tags": [
          "billing"
        ],
        "operationId": "billing.history.list",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillingHistoryListResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "SdkWorkApiResponse": {
        "type": "object",
        "required": [
          "code",
          "data",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "enum": [
              0
            ]
          },
          "data": {
            "type": "object"
          },
          "traceId": {
            "type": "string"
          }
        }
      },
      "ProblemDetail": {
        "type": "object",
        "required": [
          "code",
          "traceId",
          "title",
          "detail"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "traceId": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          }
        }
      },
...
      "BillingHistoryListResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SdkWorkApiResponse"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "required": [
                  "items",
                  "pageInfo"
                ],
                "properties": {
                  "items": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/BillingHistoryItem"
                    }
                  },
                  "pageInfo": {
                    "$ref": "#/components/schemas/PageInfo"
                  }
                }
              }
            }
          }
        ]
      },
...
      "BillingHistoryItem": {
        "type": "object",
        "required": [
          "id",
          "historyNo",
          "type",
          "direction",
          "assetType",
          "amount",
          "pointsDelta",
          "status",
          "title",
          "sourceType",
          "sourceId",
          "occurredAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "historyNo": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "direction": {
            "type": "string"
          },
          "assetType": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "currencyCode": {
            "type": "string"
          },
          "pointsDelta": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "referenceNo": {
            "type": "string"
          },
          "sourceType": {
            "type": "string"
          },
          "sourceId": {
            "type": "string"
          },
          "relatedOrderNo": {
            "type": "string"
          },
          "paymentMethod": {
            "type": "string"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PageInfo": {
        "type": "object",
        "required": [
          "mode"
        ],
        "properties": {
          "mode": {
            "type": "string",
            "enum": [
              "offset",
              "cursor"
            ]
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalItems": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "nextCursor": {
            "type": "string"
          },
          "hasMore": {
            "type": "boolean"
          }
        }
      },
...
    },
    "securitySchemes": {
      "AuthToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "SDKWork auth token carried as Authorization: Bearer <auth_token>."
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token",
        "description": "SDKWork access isolation token."
      }
    }
  },
  "security": [
    {
      "AuthToken": [],
      "AccessToken": []
    }
  ]
}